Added <DerivedFrom> attribute for Model fields
This lets you mark the data dependencies on computed fields of a model.
This hijacks invocations to the setters on the abstract properties of the
models and checks to see if there are any computed fields that depend on
the property being updated. Any children found in this process have
NotifyPropertyChanged called on the model using the child name.